chore: Rename fetch.stories.tsx to useFetch.stories.tsx
